Use HOMEBREW_ACCEPT_EULA as env variable
Changes in Homebrew from v1.4.0 onwards sanitise the environment that is
available to formulae.  Only specific env variables are passed through
and simplest way of getting the option through is to prefix the
environment variable with `HOMEBREW_`.

I have a PR open with the core of Homebrew to explain the env variable
filtering in detail but you can read the details in the meantime in the
PR at
https://github.com/Homebrew/brew/blob/7f8d90740a228a582330df684518274712ec04c9/docs/Formula-Cookbook.md#using-environment-variables